it was creates the 20 txt files to input test cases "in01.txt... in20.txt".
the filed have the extension .txt and the name is for example  in01.txt.

A maximum of 20 files (in01.txt .... in20.txt).

Every file contain 3 chains and every chain have 7 characters, For
example AAAAIAA AAAADAA DDDAAID.

The allowed characters are:A (represents and advance), D(represents a 90
degree to the rigth), L(represents a 90 degree to the left).