FileManagementTest
There are 4 test cases in this class:

TEST ONE:
The test case is to verify the folder where the input files were
storage.
The directory is "InputFiles".


TEST TWO:
The test case is to verify if all 20 input files exist in the correct.

Location, with the correct structure and the correct extension .txt

TEST THREE:
The test case is to verify the folder where the output files were
storage .
The directory is "OutputFiles".


TEST FOUR
The test case is to verify if all 20 output files were created with the
correct structure and the correct extension .txt